+++ usr/local/lib/steamos_diy/helpers/jupiter-dock-updater.py
@@ -3,7 +3,7 @@
 """
 # =============================================================================
 # PROJECT:      SteamMachine-DIY - Jupiter Dock Updater Shim
-# VERSION:      2.1.3
+# VERSION:      2.1.4
 # DESCRIPTION:  Self-sufficient shim for SteamOS Dock firmware updates.
 # PHILOSOPHY:   KISS (Keep It Simple, Stupid)
 # REPOSITORY:   https://github.com/dlucca1986/SteamMachine-DIY